move authentication to integration test setup
