--- internal/driver/driver_test.go
@@ -1,121 +1,0 @@
-// Copyright 2021 FerretDB Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package driver
-
-import (
-	"testing"
-
-	"github.com/FerretDB/wire"
-	"github.com/FerretDB/wire/wirebson"
-	"github.com/FerretDB/wire/wireclient"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	"github.com/FerretDB/FerretDB/internal/util/must"
-	"github.com/FerretDB/FerretDB/internal/util/testutil"
-)
-
-func TestDriver(t *testing.T) {
-	if testing.Short() {
-		t.Skip("skipping in -short mode")
-	}
-
-	ctx := testutil.Ctx(t)
-
-	c := must.NotFail(wireclient.Connect(ctx, "mongodb://127.0.0.1:47017/", testutil.Logger(t)))
-	t.Cleanup(func() { require.NoError(t, c.Close()) })
-
-	dbName := testutil.DatabaseName(t)
-
-	doc1 := must.NotFail(wirebson.NewDocument("_id", int32(0), "w", int32(2), "v", int32(1)))
-	doc2 := must.NotFail(wirebson.NewDocument("_id", int32(1), "v", int32(2)))
-	doc3 := must.NotFail(wirebson.NewDocument("_id", int32(2), "v", int32(3)))
-
-	// TODO https://github.com/FerretDB/FerretDB/issues/4448
-	// must.NoError(c.Authenticate(ctx))
-
-	t.Run("Drop", func(t *testing.T) {
-		dropCmd := must.NotFail(wirebson.NewDocument(
-			"dropDatabase", int32(1),
-			"$db", dbName,
-		))
-
-		resHeader, resBody, err := c.Request(ctx, must.NotFail(wire.NewOpMsg(dropCmd)))
-		require.NoError(t, err)
-		assert.NotZero(t, resHeader.RequestID)
-
-		resMsg, err := must.NotFail(resBody.(*wire.OpMsg).RawDocument()).Decode()
-		require.NoError(t, err)
-
-		ok := resMsg.Get("ok").(float64)
-		require.Equal(t, float64(1), ok)
-	})
-
-	t.Run("Insert", func(t *testing.T) {
-		insertCmd := must.NotFail(wirebson.NewDocument(
-			"insert", "values",
-			"documents", must.NotFail(wirebson.NewArray(doc1, doc2, doc3)),
-			"$db", dbName,
-		))
-
-		resHeader, resBody, err := c.Request(ctx, must.NotFail(wire.NewOpMsg(insertCmd)))
-		require.NoError(t, err)
-		assert.NotZero(t, resHeader.RequestID)
-
-		resMsg, err := must.NotFail(resBody.(*wire.OpMsg).RawDocument()).Decode()
-		require.NoError(t, err)
-
-		ok := resMsg.Get("ok").(float64)
-		require.Equal(t, float64(1), ok)
-
-		n := resMsg.Get("n").(int32)
-		require.Equal(t, int32(3), n)
-	})
-
-	var cursorID int64
-
-	t.Run("Find", func(t *testing.T) {
-		findCmd := must.NotFail(wirebson.NewDocument(
-			"find", "values",
-			"filter", must.NotFail(wirebson.NewDocument()),
-			"sort", must.NotFail(wirebson.NewDocument("_id", int32(1))),
-			"batchSize", int32(1),
-			"$db", dbName,
-		))
-
-		resHeader, resBody, err := c.Request(ctx, must.NotFail(wire.NewOpMsg(findCmd)))
-		require.NoError(t, err)
-		assert.NotZero(t, resHeader.RequestID)
-
-		resMsg, err := must.NotFail(resBody.(*wire.OpMsg).RawDocument()).Decode()
-		require.NoError(t, err)
-
-		cursor, err := resMsg.Get("cursor").(wirebson.AnyDocument).Decode()
-		require.NoError(t, err)
-
-		firstBatch, err := cursor.Get("firstBatch").(wirebson.AnyArray).Decode()
-		require.NoError(t, err)
-		cursorID = cursor.Get("id").(int64)
-		require.NotZero(t, cursorID)
-
-		arr, err := firstBatch.Decode()
-		require.NoError(t, err)
-		require.Equal(t, 1, arr.Len())
-
-		d, err := arr.Get(0).(wirebson.AnyDocument).Decode()
-		require.NoError(t, err)
-		require.Equal(t, doc1, d)
-	})
-}
